+++ packages/angular_devkit/build_angular/src/builders/dev-server/output.ts
@@ -1,0 +1,18 @@
+/**
+ * @license
+ * Copyright Google LLC All Rights Reserved.
+ *
+ * Use of this source code is governed by an MIT-style license that can be
+ * found in the LICENSE file at https://angular.io/license
+ */
+
+import { BuilderOutput } from '@angular-devkit/architect';
+
+/**
+ * @experimental Direct usage of this type is considered experimental.
+ */
+export interface DevServerBuilderOutput extends BuilderOutput {
+  baseUrl: string;
+  port?: number;
+  address?: string;
+}
